+++ core/libs/dimg/loaders/dimgloader.h
@@ -116,7 +116,7 @@ protected:
     bool                    imageHasAlpha()                                         const;
     bool                    imageSixteenBit()                                       const;
 
-    unsigned int            imageNumBytes()                                         const;
+    quint64                 imageNumBytes()                                         const;
     int                     imageBitsDepth()                                        const;
     int                     imageBytesDepth()                                       const;
